GPIO update, port works

+++ targets/arm/mikroe/renesas/include/hal_ll_target_names.h
@@ -55,10 +55,10 @@ extern "C"{
 #include <stdlib.h>
 
 typedef uintptr_t handle_t;
-typedef uint8_t hal_ll_gpio_mask_t;
+typedef uint16_t hal_ll_gpio_mask_t;
 typedef uint8_t hal_ll_pin_name_t;
 typedef uint8_t hal_ll_port_name_t;
-typedef uint8_t hal_ll_port_size_t;
+typedef uint16_t hal_ll_port_size_t;
 typedef uint32_t hal_ll_base_addr_t;
 typedef uint8_t hal_ll_channel_t;
